Remove defunct methods and tests, activate all tests
Remove remove_replicate and prune_replicates methods now that
replicates are handled by iRODS' resources. Remove corresponding
tests.

Simplify replicate counting tests. The test for the invalid_replicates
method no longer covers the case where invalid replicates exist. They
are now much harder to generate in iRODS, so we'd need to mock iRODS
for this case.

Fix the check for tests needing admin rights.

Remove TEST_AUTHOR because 1) tests always need an iRODS server 2) we
are the authors, so we always want all the tests to run.

Add iRODS 4.2.10 and baton 3.1.0 to the test matrix.