Fix failure due to a fragile test
The test failed when using Docker/Singularity wrappers because
although the test outcome is correct (the constructor raises an
error), the error message string did not match the expected pattern.

The error message differs between macOS and Linux and between
Singularity and Docker. It also depends on the shell environment.

This changes the test to simply test for failure when a failure
condition is set up.