Consolidated all logic about study-related ACLs.
Consolidated all logic about study-related ACLs
in the expected_groups method of WTSI::NPG::iRODS::Path thus
making this part of logic independent of the sequencing platform.
In future this logic can be applied, if necessary, to iRODS
collections.

Elements of this logic are moved from WTSI::NPG::iRODS::DataObject.
The logic related to xahuman is currently inplemented in npg_irods
GitHub package.